Recursively infer whether methods/functions/closures are pure.
Build a graph, infer the function is impure if it depends on any
functions/methods that aren't pure.
