+++ tests/misc/empty_methods_plugin_test/src/0000_empty_methods_functions.php
@@ -1,7 +1,7 @@
 <?php
 
 class ClassWithEmptyMethods756 {
-    public function one() {} // no warning, overriden
+    public function one() {} // no warning, overridden
 
     public function two() {}
