+++ tests/php74_files/expected/018_arrow_func_call_use_retval.php.expected
@@ -1,0 +1,1 @@
+%s:5 PhanPluginUseReturnValueKnown Expected to use the return value of the user-defined function/method \double18
