+++ tests/php74_files/src/018_arrow_func_call_use_retval.php
@@ -1,0 +1,5 @@
+<?php
+function double18(array $args) : array {
+    return array_map(fn ($x) => $x * 2, $args);
+}
+double18([2,3]);
